FOUR-8401 Removed old controls rail * Moved the rankConstants file to the explorer-rail
